Add IAITask interface and corresponding unit tests
Added a new interface `IAITask` in the `AIServiceLib.Core.Interfaces` namespace with a method `RunAsync` that returns a `Task`. Created a new test class `InterfaceTests` in the `AIServiceLib.Tests` namespace. Added a mock implementation of the `IAITask` interface named `MockAiTask` within the `InterfaceTests` class. Implemented a test method `IAITaskTest` in the `InterfaceTests` class to verify that an instance of `MockAiTask` can be created and is not null.